requirements.txt
# Function dependencies, for example:
# package>=version
requests
slack_sdk
google-cloud-firestore
pytz
datetime
python-dotenv
